requirements.txt
Pillow
blobfile
mypy
numpy
pytest
requests
torch
torchvision
dall-e